+++ spec/kettle/soup/cover/rakelib/turbo_tests_spec.rb
@@ -3,6 +3,7 @@
 require "kettle/soup/cover/tasks"
 
 # rubocop:disable RSpec/DescribeClass
+# rubocop:disable RSpec/MultipleDescribes
 RSpec.describe "rake turbo_tests:cleanup" do
   let(:rake) { Rake::Application.new }
   let(:gem_root) { File.expand_path("../../../../../..", __FILE__) }
@@ -42,5 +43,15 @@ RSpec.describe "rake turbo_tests:setup" do
 
     expect(Kettle::Soup::Cover).to have_received(:clear_turbo_tests_coverage_dir!)
   end
+
+  it "does not clear turbo_tests2 coverage when coverage is disabled" do
+    allow(Kettle::Soup::Cover).to receive(:turbo_tests_coverage?).and_return(false)
+    allow(Kettle::Soup::Cover).to receive(:clear_turbo_tests_coverage_dir!)
+
+    rake["turbo_tests:setup"].invoke
+
+    expect(Kettle::Soup::Cover).not_to have_received(:clear_turbo_tests_coverage_dir!)
+  end
 end
+# rubocop:enable RSpec/MultipleDescribes
 # rubocop:enable RSpec/DescribeClass